+++ language-server/src/tests/action.rs
@@ -13817,6 +13817,24 @@ fn op(i: Int) -> Int {
 }
 
 #[test]
+fn unwrap_anonymous_function_can_only_trigger_if_cursor_is_within_the_function() {
+    assert_no_code_actions!(
+        UNWRAP_ANONYMOUS_FUNCTION,
+        "import gleam/list
+
+pub fn main() {
+  list.map([1, 2, 3], fn(int) { op(int) })
+}
+
+fn op(i: Int) -> Int {
+  todo
+}
+",
+        find_position_of("fn(int)").select_until(find_position_of("todo"))
+    );
+}
+
+#[test]
 fn unwrap_trivial_anonymous_function_with_bad_spacing() {
     assert_code_action!(
         UNWRAP_ANONYMOUS_FUNCTION,